+++ test/src/utils/file.h
@@ -25,8 +25,28 @@ int test_access_dir(const char* path);
 
 /**
  * @brief Path to current executable.
+ * @return Path.
  */
 const char* test_get_self_exe(void);
+
+/**
+ * @brief Path to current executable directory.
+ * @return Path
+ */
+const char* test_get_self_dir(void);
+
+/**
+ * @brief Get current executable name.
+ * @return Executable name.
+ */
+const char* test_self_exe_name(void);
+
+/**
+ * @brief Get file name from full path.
+ * @param file Full file path.
+ * @return file name.
+ */
+const char* test_file_name(const char* file);
 
 #ifdef __cplusplus
 }